requirements.txt
Django==1.6.1
South==0.8.4
django-authtools==0.2.1
django-cachebuster==0.2.1
django-extensions==1.3.3
django-floppyforms==1.1.1
ipython==1.1.0
pygithub3==0.5
readline==6.2.4.1
requests==0.14.0
six==1.5.2
wsgiref==0.1.2
